fix: deleeteCarreras

+++ src/controller/CarrerasControllers.js
@@ -1,7 +1,9 @@
 import Carreras from "../models/Carerra.js";
+import Materias from "../models/Materias.js";
+
 
 class carrerasController {
-  
+
   static createCarreras = async (req, res) => {
     const carrera = new Carreras(req.body);
     try {
@@ -31,10 +33,20 @@ class carrerasController {
   static deleteCarrera = async (req, res) => {
     const { id } = req.params;
     try {
-      const carrera = await Carreras.findByIdAndDelete(id);
+
+      const carrera = await Carreras.find(id)
       if (!carrera) {
         return res.status(404).send('Carrera no encontrada');
       }
+
+      //verifico que la carrera no tenga materias asociadas
+      const materiasAsociadas = await Materias.find({ 'carreras': id })
+      if (materiasAsociadas > 0) {
+        return res.status(404).send('No se puede eliminar la carrera. Hay materias asociadas.')
+      }
+
+      await Carreras.findByIdAndDelete(id);
+
       res.send('Carrera eliminada correctamente');
     } catch (error) {
       console.log(error);
